Rename package according to #SPONGE-35
